backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
PyMuPDF>=1.26.0
openai>=1.35.5
sqlalchemy>=2.0.31
asyncpg>=0.29.0
alembic>=1.13.2
razorpay>=1.4.1
reportlab>=4.2.2
python-dotenv>=1.0.1
pydantic>=2.8.2
httpx>=0.27.0
aiosqlite>=0.20.0

# Testing
pytest>=8.2.2
pytest-asyncio>=0.23.7
